[bazel][fidl] Define an IDK atom when appropriate
Move the JSON summary file to the main build directory to comply with
symbolic macro naming requirements, and remove `gen_dir` from `fidlc()`
since it is unused.

Define //sdk/fidl, which contains atoms for FIDL libraries in the IDK.

Migrate fuchsia.hardware.spi.businfo and fuchsia.math to Bazel as
examples of unstable and stable, respectively, libraries in the IDK.

Always place `.fidl` files in the root of the library's directory in the
IDK to simplify logic. Add an assert to the GN template to ensure this
is not a change in behavior.

Bug: 428285014
Test: fx bazel build --config=fuchsia //sdk/fidl
Change-Id: I1b8ece1f531d119ab45ee29b0b8fdc18ab3e214b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1444484
Commit-Queue: David Dorwin <ddorwin@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>