[sestarnix] Update the compile_policies.py script checkpolicy arg
Remove the --checkpolicy-executable argument and instead require
that a checkpolicy binary has been placed in the "local" sub-
directory of the Fuchsia checkout.

In future the scripts will be updated to use a prebuilt checkpolicy
rolled from infra, and the logic migrated to a set of build rules.

Bug: 324115065
Change-Id: Ib3f1efcb96d334c937b99f95107469c461b5906a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1452269
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Wez <wez@google.com>
Reviewed-by: Ambre Williams <ambre@google.com>