[bazel][fidl] Fix library dependencies
Pass the fidlc target for dependency libraries to the `fidlc()` rule
because these are the targets that have the necessary provider. Also
record the source files for all dependency libraries, including indirect
dependencies, and include these in `inputs` to the `fidlc` action so
that all necessary files are in the sandbox.

In addition, make the non-IDK FIDL targets depend on the corresponding
targets for all dependency FIDL libraries.

The Test commands below test direct dependencies in an IDK FIDL library
and indirect dependencies in a FIDL library, respectively. These will
only work after the commits migrating these libraries land.

Bug: 428285014
Test: fx bazel build --config=fuchsia //sdk/fidl/fuchsia.images2:fuchsia.images2_idk
Test: fx bazel build --config=fuchsia //sdk/fidl/fuchsia.hardware.display.engine
Change-Id: I2bb43d84bea812def12811f98de1defcad664963
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1448075
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>