[oom-watcher] Migrate to Shutdown
This is a behavioral no-op.

PerformReboot, RebootToBootloader, RebootToRecovery, and Poweroff are
deprecated in favor of Shutdown. Compared to the deprecated methods,
Shutdown allows ALL clients to specify the reason for the shutdown.
Previously, only PerformReboot allowed clients to specify the shutdown
reason.

Feedback will use the additional information to improve the
determination of its canonical reboot reason, including but not limited
to filing crash reports for high temperature poweroffs.

Bug: 465530639
Test: `fx test //src/bringup/bin/critical-services`
Test: `fx shell k mem oom`
Change-Id: I6e16223bab17ebf3e1f0f8189f92f4be4e8e2ed3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1451556
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Thomas Gales <tgales@google.com>