[bt][gap] Commit device class to persisted BondingData
Add the device class field to the general BondingData struct defined in
the fuchsia-bluetooth crate. Also add an optional inspect node for this
property if it is set.

Update bt-gap to save and load this field so that it is persisted across
reboots.

Note that in practice, the device class will continue to not be present
in sys.Access/WatchPeers updates from the host stack as bt-host is not
updated to parse & save this field.

Bug: 468338802
Test: fx test bt-gap-unittests

Change-Id: I6b6488dc612325e45d6e639c3383f505133631c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1448211
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>