[bt][avrcp] Rework connection logic to be on-demand
Previously, the AVRCP profile would automatically attempt to establish
control and browse connections as soon as a peer's services were
discovered. This is inefficient and has been reworked to an on-demand
model.

This change modifies the peer connection logic to only initiate AVCTP
control and browse connections when `GetControllerForTarget` or
`GetBrowseControllerForTarget` is explicitly called by a client.

`request_control_connection` and `request_browse_connection`
are exposed on the `RemotePeerHandle` to trigger connection attempts
from the FIDL service layer when a controller is requested.

Test: fx test bt-avrcp-tests
Bug: 450144120

Change-Id: Iea7045ec33ea7afdc21aee12b3844058f84663fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1445968
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Dayeong Lee <dayeonglee@google.com>