[cog] Pass repository root from shell script to git.py
The repository root is now determined in the shell script and passed to git.py via a new --repository-root argument. This simplifies the Python script by removing the need to find the root itself. A check is added in git.py to ensure the provided repository root is within a cog workspace.

Change-Id: Ie9453adb6a9e843e1e59c3c5d26d521420fa4417
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1453270
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>