[kernel][arm] honor the IDC and DIC bits in CTR_EL0 when mapping executable pages
With newer processors, less work can be done when synchronizing I and D
caches after mapping or unmapping a page. Honor the bits in this
register which tells you what kind of I cache the cpu has and what you
need to do to synchronize it.

Change-Id: I8a2c45d24648368aa3057a13d9b286db76579b6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1447109
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>