[fidl] Add create_proxy_and_stream to ZirconClient
ZirconClient is just a zero-sized struct that can be a receiver for
various utility methods that we'd normally call as receiverless
functions. This lets us have a common API between normal and FDomain
handles, which usually require a receiver as they refer to a
connection.

This patch just adds one more method to ZirconClient that was missing.

Change-Id: I6a6a696478f458873982e50774404425cac7bf3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1452815
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>