[feedback] Make report store size configurable
The report store persists all non-snapshot data if connectivity is
unavailable, including minidumps. For some processes, such as the
Starnix kernel, minidumps can be quite large and may exceed Feedback's
report store limit of 500 KiB for /cache.

This CL adds a flag to Feedback's config so that assembly can specify
the size of the report store. This CL does NOT change the current size
of the store. A follow-up CL will increase the store size for products
that specify the "large_disk" Feedback assembly flag as "true."

Bug: 465548680
Test: `fx test //src/developer/forensics`
Test: `fx test
//src/lib/assembly/platform_configuration:platform_configuration_test`

Change-Id: Idc6e0e2762b818acd9e6bcf773b30a56491dea8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1444818
Commit-Queue: Thomas Gales <tgales@google.com>
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>