[wlan] Use Constant for channel reporting
Reporting channel even when we are disconnected from a network makes the
visualization of this data confusing, since the other data (phy rate,
RSSI) interpolates 0 when we are not actively connected to a network.
Using a constant fill for channel data means that we will only see data
reported when we are connected to a network.

Fixed: 466071471
Change-Id: I134c7ead46961b5369bd9c4e54e3892e2d5696f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441519
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>