[sdk] Add strict audit script.
This is a vibe-coded tool that identifies all the `strict` items in the
SDK surface, and tags them with the `strict_audit` attribute, indicating
that they should be audited for evolvability. See
https://fxbug.dev/42055534 for more details.

This is probably a one-time-use script, but I dunno, it feels better to
check it in than to not. The next CL in the chain will actually run the
thing and add the annotations.

Note that this also adds the "locations" to fields in the platform IR.
It brings the resulting IR file from like ~20MB to ~30MB, but I don't
think that's a big deal.

Bug: 42055534

Change-Id: Ib84dfec7a19fef8e1b565e585fe62a59c1c9ddde
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1444509
Reviewed-by: Ian McKellar <ianloic@google.com>
Fuchsia-Auto-Submit: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>