[recovery] Migrate to Shutdown
This is a behavioral no-op.

PerformReboot, RebootToBootloader, RebootToRecovery, and Poweroff are
deprecated in favor of Shutdown. Compared to the deprecated methods,
Shutdown allows ALL clients to specify the reason for the shutdown.
Previously, only PerformReboot allowed clients to specify the shutdown
reason.

Feedback will use the additional information to improve the
determination of its canonical reboot reason, including but not limited
to filing crash reports for high temperature poweroffs.

Bug: 465530639
Test: `fx test //src/recovery/lib/recovery-util`
Change-Id: I351db4da7b525f32035260d4e9d27c204ac3a4b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1456064
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Commit-Queue: Thomas Gales <tgales@google.com>