[kernel][ktl] Soft transition for libc++ std::align change
This is the first half of a soft transition for ktl to work with
the libc++ header implementation of std::align -> ktl::align.

This changes the out-of-line definition to use manual C++ name
mangling as a means to hide the definition's true name from the
compiler.  This doesn't change what gets defined at all.

Once the new libc++ rolls in, its header will have an inline
definition for std::align, so that definition will not be used
any more and can be removed.

Bug: 470118858
Change-Id: Ib157194451b2fdcd60f5f5a7b6bd4add74e16310
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1456168
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Paul Kirth <paulkirth@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>