[ffx] Shorten the description of the options and commands
This cleans up the descriptions of main commands and options so that
they fit in a single line in the terminal.

Thus, when the user types `ffx` the visual output is much cleaner and
easier to read. In most cases, especially for the description of the
options, it is more concise and more obvious on what it does.

Resulting CLI golden files are also updated.

Before: https://paste.googleplex.com/4985969088528384?raw
After: https://paste.googleplex.com/5801500217573376?raw

Test: Run `ffx`
Change-Id: I1ff0c35db85f154c6ed342c90763217fd7861d50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1458173
Commit-Queue: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>