[kernel][usercopy] copy_array_* remove duplication
The copy_array_* functions in usercopy have a variants that either copy
from the start of the passed array, or from a provided offset. The code
is almost identical, so duplication can be removed.

This is applied to all 4 of the copy_array_* functions, copy to and from
user and with or without capture faults.

Change-Id: I9cf3af7947b45980aac71cce636c58cadea961e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1448848
Commit-Queue: Sage Barreda <sagebarreda@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>