[ffx] Clean up emulator error message during compatibility error
During the emulator lifecycle, we get a lot of messages in the info logs
that could be more useful and actionable to the user. Move some of those
messages to debug and add more context to others.

As an example, this was an error during emulator launch that was not
helpful or actionable:

stdout:
```
Waiting for Fuchsia to start (up to 60 seconds).............
Emulator is ready.
Compatibility status: Error Host connection failed: creating pipe to fuchsia-emulator failed: could not establish connection: "Connection timed out during banner exchange\r\n"```
```

With this fix, it tells the user to attempt to restart the emulator.

Test: ffx emu start && ffx target list && ffx target show
Change-Id: Idbcb0d16dcfef678c02891bcbb6700b9ce1372af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1458323
Commit-Queue: Amit Uttamchandani <amituttam@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>