[bazel][boringssl] Reduce copy span of @boringssl repository.
For historical reasons, the @boringssl repository must currently be
populated by hard-linking the content of //third_party/boringssl
into the repository directory, and creating a patched BUILD.bazel
file for it. That source directory is huge though.

This CL reduces the amount of files being copied, from 569 MiB
to 139 MiB, by omitting a few sub-directories that are not needed
by the build.

It also improves the repository rule's tracking of input files a
bit, ensuring it will be run when all its inputs are modified
(the previous version ignored modifications to the hardlink script
and the content hash file, in particular!).

This is only a temporary work-around until a better solution, such
as https://fxrev.dev/1442608, which completely gets rid of the
copy, can be submitted to the tree.

Change-Id: Ia414188ff18c0b2dee2f1de41b1989a466cdfd5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1458728
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Turner <digit@google.com>