[ffx] Add flags for tweaking FDomain interactions
If you call read() on an FDomain socket, it sends a message to the
target, and then a reply comes back with data. Doing that several times
means several round trips. For users that want to avoid that we have an
FDomain-specific streaming API that tells the target to send data
eagerly rather than waiting for us to request it. That seems more
appropriate for a speed test so this CL switches to it, and adds a flag
to go back to the conventional API if desired.

Writes are similar. Every write incurs a round trip to the target. The
solution here is simple: don't await every write. Call write several
times, then await each returned future to check for errors. We
similarly switch to this method, and add a flag to tweak how many
writes can be done in parallel at once. Setting to 1 restores the old
behavior.

Change-Id: I6a6a6964af75cc8b92d3119ac1ec48897743ba07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1455140
Commit-Queue: Casey Dahlin <sadmac@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>